requirements.txt
asgiref==3.8.1
attrs==23.2.0
Django
django-cors-headers==4.3.1
django-userforeignkey==0.5.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-extra-fields==3.7.0
drf-spectacular==0.27.2
enum34==1.1.10
filetype==1.2.0
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
sqlparse==0.4.4
typing_extensions==4.11.0
uritemplate==4.1.1
